urlshortner/requirements.txt
Django==1.11.6
django-rest-framework==0.1.0
djangorestframework==3.7.0
pytz==2017.2
